requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
# Django==2.1.1
django-allauth==0.38.0
django-audiofield==0.10.0
django-celery==3.2.2
django-crispy-forms==1.7.2
django-datatables-view==1.17.0
django-jet==1.0.7
django-markitup==3.0.0
django-tinymce==2.7.0
djangorestframework==3.8.2
idna==2.7
kombu==4.2.1
Markdown==2.6.5
mysqlclient==1.3.12
oauthlib==2.1.0
Pillow==5.3.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
vine==1.1.4
